+++ src/main/java/racingcar/model/CarGroup.java
@@ -1,4 +1,6 @@
-package racingcar;
+package racingcar.model;
+
+import racingcar.model.Car;
 
 import java.util.*;
