+++ src/main/java/racingcar/view/Message.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.view;
 
 public class Message {
